add headers to tests
