+++ packages/chatbot-server-mongodb-public/src/index.ts
@@ -8,8 +8,6 @@ import {
   CORE_ENV_VARS,
   assertEnvVars,
 } from "mongodb-chatbot-server";
-import https, { request, RequestOptions } from "https";
-import http from "http";
 import { URL } from "url";
 
 export const {
@@ -25,6 +23,7 @@ export const {
   NODE_ENV,
 } = assertEnvVars(CORE_ENV_VARS);
 import { config, mongodb, embeddedContentStore } from "./config";
+import { addHeadersToHttpsRequests } from "./addHeadersToHttpRequests";
 
 const PORT = process.env.PORT || 3000;
 
@@ -64,55 +63,4 @@ try {
 } catch (e) {
   logger.error(`Fatal error: ${e}`);
   process.exit(1);
-}
-
-/**
-  This function appends arbitrary headers to HTTPS requests made to a specific hostname.
- */
-function addHeadersToHttpsRequests(
-  targetHostname: string,
-  headers: Record<string, string>
-) {
-  const originalHttpsRequest = request;
-
-  function requestWrapper(originalRequest: typeof request): typeof request {
-    const wrappedRequest = function (
-      arg1: string | RequestOptions | URL,
-      arg2?: RequestOptions | ((res: http.IncomingMessage) => void),
-      arg3?: (res: http.IncomingMessage) => void
-    ) {
-      let options: RequestOptions | undefined;
-
-      if (typeof arg1 === "string" || arg1 instanceof URL) {
-        if (typeof arg2 === "function") {
-          options = {};
-        } else {
-          options = arg2 || {};
-        }
-      } else {
-        options = arg1;
-      }
-
-      // Extract hostname from options
-      const hostname =
-        options.hostname ||
-        (typeof arg1 === "string" ? new URL(arg1).hostname : arg1.hostname);
-
-      // Check if the hostname is 'targetHostname' and modify headers
-      if (hostname === targetHostname) {
-        logger.log("Adding headers for Request to:", options.path);
-        options.headers = options.headers || {};
-        for (const [key, value] of Object.entries(headers)) {
-          options.headers[key] = value;
-        }
-      }
-
-      // Call the original request function with correct parameters
-      return originalRequest(arg1 as any, arg2 as any, arg3);
-    };
-
-    return wrappedRequest;
-  }
-
-  https.request = requestWrapper(originalHttpsRequest);
 }
